+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Changed base of shapes to `compas_viewer.scene.ShapeObject`.
 * Changed `FrameObject` to something with just 3 axes.
 * Restructured `SceneObjects` to accepct `item` as kwargs.
+* Fixed `compas_viewer.components.slider` step attribute.
 
 ### Removed